requirements.txt
streamlit>=1.30,<2.0
fasttext-wheel>=0.9.2
pandas>=2.0,<2.3
numpy<2